Apply review comments

+++ src/cpp/include/openvino/genai/generation_config.hpp
@@ -76,7 +76,7 @@ public:
     // Multinomial
     float temperature = 1.0f;
     float top_p = 1.0f;
-    int top_k = 50;
+    size_t top_k = 50;
     bool do_sample = false;
     float repetition_penalty = 1.0f;
